Extract base configuration
