+++ src/main/scala/intellij/haskell/runconfig/HaskellConsoleConfigurationType.scala
@@ -3,14 +3,14 @@ package intellij.haskell.runconfig
 import com.intellij.execution.configurations.{ConfigurationFactory, ConfigurationType}
 import intellij.haskell.HaskellIcons
 
-final class HaskellConsoleConfigurationType extends ConfigurationType {
-  def getDisplayName = "Haskell Stack Shell"
+final class HaskellStackConfigurationType extends ConfigurationType {
+  def getDisplayName = "Haskell Stack"
 
-  def getConfigurationTypeDescription = "Haskell Stack Shell configuration"
+  def getConfigurationTypeDescription = "Haskell Stack configuration"
 
   def getIcon = HaskellIcons.HaskellSmallLogo
 
-  def getId = "HaskellConsoleConfigurationType"
+  def getId = "HaskellStackConfigurationType"
 
   def getConfigurationFactories: Array[ConfigurationFactory] = Array[ConfigurationFactory](new HaskellConsoleConfigurationFactory(this))
 }
